+++ src/main.c
@@ -11,10 +11,10 @@
 
 int main () {
 
-	uint exit = 0;
-	uint state = 0; /* 0-intro,1-history,2-game */
-	uint grapset = 0; /* 0-8bits, 1-16bits */
-	uint fullscreen = 0; /* 0-Windowed,1-Fullscreen */
+	uint8_t exit = 0;
+	uint8_t state = 0; /* 0-intro,1-history,2-game,3-gameover,4-ending,5-exit */
+	uint8_t grapset = 0; /* 0-8bits, 1-16bits */
+	uint8_t fullscreen = 0; /* 0-Windowed,1-Fullscreen */
 
 	/* SDL2 initialization */
 	SDL_Init(SDL_INIT_VIDEO | SDL_INIT_AUDIO);
