+++ src/startscreen.c
@@ -2,7 +2,7 @@
 
 # include "base.h"
 
-void startscreen(SDL_Window *screen,uint *state,uint *grapset,uint *fullscreen) {
+void startscreen(SDL_Window *screen,uint8_t *state,uint8_t *grapset,uint8_t *fullscreen) {
 
 	uint8_t exit = 0;
 	uint8_t musicplay = 0;
